sbxxx: Add spispeed parameter.
Allow to set the SPI clock frequency on SB600 et al. with a programmer
parameter. The default is to not change the preset value. If the parameter
is given (and matches a possible value), fast reads are disabled and the
SPI clock is set temporarily. Both registers are restored on programmer
shutdown.

Example: ./flashrom -p internal:spispeed=33 -V

Possible values for spispeed are "66/reserved", "33", "22", "16.5".

Signed-off-by: Stefan Tauner <stefan.tauner@student.tuwien.ac.at>